# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Mastori Api specification",
        "description": "",
        "contact": {
            "email": "giannaropoulou.foteini@gmail.com;tania.pets@gmail.com"
        },
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/professions": {
            "get": {
                "tags": [
                    "professions"
                ],
                "description": "Returns all the professions",
                "operationId": "getprofessions",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All professions response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/profession"
                            }
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "professions"
                ],
                "description": "Adds a new profession in database",
                "operationId": "addProfession",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profession"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns created profession object",
                        "schema": {
                            "$ref": "#/definitions/profession"
                        }
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid/Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
        "/professions/{id}": {
            "put": {
                "tags": [
                    "professions"
                ],
                "description": "Edits a professions",
                "operationId": "editprofession",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of profession to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/profession"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Professions response",
                        "schema": {
                            "$ref": "#/definitions/profession"
                        }
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    },
                    "400": {
                        "description": "validation errors",
                        "schema": {
                            "$ref": "#/definitions/validationsErrorsModel"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "professions"
                ],
                "description": "Delete a profession",
                "operationId": "deleteprofession",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of profession to delete",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "token_not_provided/token_invalid",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    },
                    "500": {
                        "description": "Cannot be deleted.",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errorModel": {
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "fieldValidationDescription": {
            "properties": {
                "error_description": {
                    "type": "string"
                }
            }
        },
        "validationerror": {
            "properties": {
                "fieldName": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fieldValidationDescription"
                    }
                }
            }
        },
        "validationsErrorsModel": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validationerror"
                    }
                }
            }
        },
...
        "profession": {
            "required": [
                "tag",
                "title"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "readOnly": true
                },
                "tag": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                }
            }
        },
...
    }
}